requirements.txt
ultralytics
aiohttp
aiofiles
contextly
openai-clip
geopandas
openpyxl